PYIC-905: Return state in auth response
We need to return the state value given to us in the auth request back
to the user in the auth response. This change pulls it from the frontend
session and sets it as a query param.

This also moves to using a proper URL object to parse the redirect URL
and add the query string params. This is important because if the
redirect URL configured for the client is just a host (ie doesn't
already have any search query params), the client redirect will result
in a 404. This is because we were just appending a string to the end of
the host that began with an `&`, rather than a '?'. This is the treated
as part of the path and the core front will not be able to resolve it.