PYIC-1460 Increase Desired Task Count in Production to 2
This is necessary for increased resilience by running across two
availability zones. No other changes should be necessary since Fargate's
default deployment strategy is to spread across AZs.

Update the template so that the task desired count is defined within the
template rather than passing from CI configuration. This should make it
simpler to amend and find in future and simplify the roll-out of such
changes.