PYIC-1484: Send session ID to passport back
We're implementing structured logging to make it easy to link requests
in our logs. Currently we have no shared session ID available to
passport back. There is a PR open to start logging a session ID that is
passed in from passport-front, where we currently store a users session
details. Note, this is different to core which has a session table in
the backend.

This adds the session ID of the users frontend session as a header to
the authorize lambdas.