PYIC-1490: Log passportSessionId
This adds a new field to the request and app logs, `passportSessionId`.
This will log the same session ID as the request logs currently log,
however the name will help to differentiate it from other session IDs we
may log in the future, such as a persistent session ID. The backend is
also set to log this value under this field name so the front and back
logs can be tied together easily. This naming is also consistent with
how we're logging the ipvSessionId in the core.

This also adds some light app logs, mostly as a way to demonstrate how
to use the logger. Passing in the request object is required due to the
use of the `meta` configuration block. If the request is not passed in
the logger will just log `undefined`.