BAU: Update frontend logging to use correct passportSessionId
The passportSessionId was previously the same as the frontend session. With the updates to the passport cri to enable retrying passport entry, we have updated the session model such that passport back has it's own session that is shared with passport front.

Passport front was still logging its session.id as the passportSessionId. Update to resolve this.